backend/ai-glasses-agent-backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Devin AI <158243242+devin-ai-integration[bot]@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.6"}
psycopg = {extras = ["binary"], version = "^3.2.3"}
python-multipart = "^0.0.20"
pillow = "^11.1.0"
deep-translator = "^1.11.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
